# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shipments:
    get:
      operationId: listShipments
      tags: [shipments]
      summary: List shipments
      parameters:
        - name: orderId
          in: query
          required: false
          schema:
            type: string
        - name: siteId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of shipments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shipment"
    post:
      operationId: createShipment
      tags: [shipments]
      summary: Create a shipment against a paid order; validates against remaining order quantities and creates "out" stock movements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShipmentInput"
      responses:
        "201":
          description: Created shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
        "400":
          description: Validation error (order not paid or quantity exceeds remaining order amount)
  /shipments/{id}:
    get:
      operationId: getShipment
      tags: [shipments]
      summary: Get a shipment by id (used for накладная printing)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
    patch:
      operationId: updateShipment
      tags: [shipments]
      summary: Update a shipment's metadata (note, driver, site, date). Items/quantities are immutable once shipped.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShipmentInput"
      responses:
        "200":
          description: Updated shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
  /shipments/{id}/dispatch:
    post:
      operationId: dispatchShipment
      tags: [shipments]
      summary: >
        Передать курьеру: locks the shipment against further edits and makes its
        накладная printable. Cannot be undone via the API.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dispatched shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
        "400":
          description: Shipment already dispatched
  /shipments/{id}/delete:
    post:
      operationId: deleteShipment
      tags: [shipments]
      summary: >
        Admin-only: mark a shipment as deleted (soft delete) with a mandatory note,
        instead of editing it. Reverses stock via a compensating "in" movement for
        each item. Cannot be undone via the API; create a new shipment instead.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteShipmentInput"
      responses:
        "200":
          description: Deleted (soft) shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
        "400":
          description: Shipment already marked as deleted
        "403":
          description: Only admins can mark a shipment for deletion
...
components:
  schemas:
...
    ShipmentItemInput:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: number
          exclusiveMinimum: 0
      required: [productId, quantity]
    ShipmentItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        productName:
          type: string
        productUnit:
          type: string
        quantity:
          type: number
        price:
          type: number
          description: Цена товара из соответствующей позиции заказа
      required: [id, productId, productName, productUnit, quantity, price]
    Shipment:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        orderNumber:
          type: string
        clientName:
          type: string
        siteId:
          type: string
        siteName:
          type: string
        siteAddress:
          type: string
        driverUserId:
          type: [string, "null"]
          format: uuid
        driver:
          type: string
        shipmentDate:
          type: string
          format: date
        note:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/ShipmentItem"
        totalAmount:
          type: number
        isDispatched:
          type: boolean
          description: True once "Передать курьеру" was pressed; locks editing and enables printing.
        dispatchedAt:
          type: string
          format: date-time
          nullable: true
        isDeleted:
          type: boolean
          description: >
            True once an admin marked this shipment for deletion instead of editing it.
            Stock was reversed via a compensating movement; document is read-only.
        deletedAt:
          type: string
          format: date-time
          nullable: true
        deleteNote:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - orderNumber
        - clientName
        - siteId
        - siteName
        - siteAddress
        - driverUserId
        - driver
        - shipmentDate
        - note
        - items
        - totalAmount
        - isDispatched
        - dispatchedAt
        - isDeleted
        - deletedAt
        - deleteNote
        - createdAt
    DeleteShipmentInput:
      type: object
      properties:
        note:
          type: string
          minLength: 1
      required: [note]
    CreateShipmentInput:
      type: object
      properties:
        orderId:
          type: string
        siteId:
          type: string
        driverUserId:
          type: [string, "null"]
          format: uuid
        shipmentDate:
          type: string
          format: date
        note:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/ShipmentItemInput"
          minItems: 1
      required: [orderId, siteId, shipmentDate, items]
    UpdateShipmentInput:
      type: object
      properties:
        siteId:
          type: string
        driverUserId:
          type: [string, "null"]
          format: uuid
        shipmentDate:
          type: string
          format: date
        note:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/ShipmentItemInput"
          minItems: 1
...